feat: extract @caravan/messages package for signing protocol primitives
The verifier, Entry type, and MessageSigningError taxonomy aren't
keystore-specific — they're the recovery side of any per-cosigner
message-signing flow. Pulling them into their own package now (at
@caravan/wallets's 1.0 inflection) so consumers can depend on the
verification surface without pulling the keystore drivers, and so
the future Sign-Message UI has a clean import target.

@caravan/wallets keeps the SignMessage factory, per-keystore
classes, and wrapSdkError (keystore SDK error classification stays
with the drivers). Signature fixtures live in
packages/messages/test-fixtures/signatures.json, regenerated via
`npm run generate-fixtures --workspace=@caravan/messages` when
TEST_FIXTURES changes.
