+++ .changeset/brave-pandas-sing.md
@@ -1,0 +1,5 @@
+---
+"@caravan/messages": major
+---
+
+New package. Protocol-layer primitives for signing and verifying Bitcoin messages against a known cosigner pubkey — `Entry`, `MessageSigningError`, `validateMessage`, `verifyMessageSignature` (BIP-137 / BIP-322 loose-mode), `MAX_MESSAGE_BYTES`. Pairs with the keystore drivers in `@caravan/wallets`.
