+++ apps/coordinator/package.json
@@ -98,6 +98,7 @@
     "@caravan/clients": "*",
     "@caravan/descriptors": "^0.2.0",
     "@caravan/health": "*",
+    "@caravan/messages": "*",
     "@caravan/psbt": "*",
     "@caravan/transactions": "*",
     "@caravan/typescript-config": "*",
