+++ packages/caravan-wallets/package.json
@@ -80,6 +80,7 @@
     "@babel/polyfill": "^7.7.0",
     "@babel/runtime": "^7.23.9",
     "@caravan/bitcoin": "*",
+    "@caravan/messages": "*",
     "@caravan/psbt": "*",
     "@keystonehq/bc-ur-registry": "^0.7.0",
     "@ledgerhq/hw-app-btc": "10.22.1",
@@ -87,7 +88,6 @@
     "@ledgerhq/hw-transport-webusb": "6.29.8",
     "@trezor/connect-web": "^9.6.0",
     "bignumber.js": "^8.1.1",
-    "bip322-js": "^3.0.0",
     "bitbox-api": "^0.10.0",
     "bitcoinjs-lib": "^5.1.10",
     "bowser": "^2.6.1",
