+++ packages/messages/.eslintrc.js
@@ -1,0 +1,4 @@
+module.exports = {
+  extends: ["@caravan/eslint-config/library.js"],
+  parser: "@typescript-eslint/parser",
+};
