+++ packages/messages/src/index.ts
@@ -1,0 +1,4 @@
+export type { Entry, MessageSigningErrorKind } from "./types";
+export { MessageSigningError } from "./types";
+export { MAX_MESSAGE_BYTES, validateMessage } from "./validate";
+export { verifyMessageSignature } from "./verify";
