+++ packages/messages/tsconfig.json
@@ -1,0 +1,3 @@
+{
+  "extends": "@caravan/typescript-config/base.json"
+}
