+++ packages/messages/tsup.config.js
@@ -1,0 +1,3 @@
+import { defineConfig } from "tsup";
+
+export default defineConfig({});
